Move RealmFeedImage to its own file
