+++ FeedStoreChallenge/RealmFeedStore.swift
@@ -9,25 +9,6 @@
 import Foundation
 import RealmSwift
 
-internal class RealmFeedImage: Object {
-	@objc dynamic var _id: String!
-	@objc dynamic var desc: String?
-	@objc dynamic var location: String?
-	@objc dynamic var url: String!
-		
-	internal convenience init(withLocalImage image: LocalFeedImage) {
-		self.init()
-		self._id = image.id.uuidString
-		self.desc = image.description
-		self.location = image.location
-		self.url = image.url.absoluteString
-	}
-		
-	internal var local: LocalFeedImage {
-		return LocalFeedImage(id: UUID(uuidString: _id)!, description: desc, location: location, url: URL(string: url)!)
-	}
-}
-
 public class RealmFeedStore: FeedStore {
 	
 	private var realm: Realm!
